Added in new motor wheel for an intake
I added in a new motor for an intake, this motor will be used for odometry by having an encoder in it's slot but not using the encoder for the actual motor spin. I added in run without encoder to ensure that it will not use the encoder but we can still use it.